tweak flags setup

+++ src/engine/flags.cpp
@@ -1,11 +1,11 @@
 #include "engine/flags.h"
 
-ABSL_FLAG(int, gateway_conn_per_worker, 2, "");
-ABSL_FLAG(int, sequencer_conn_per_worker, 2, "");
-ABSL_FLAG(int, shared_log_conn_per_worker, 2, "");
+ABSL_FLAG(size_t, gateway_conn_per_worker, 2, "");
+ABSL_FLAG(size_t, sequencer_conn_per_worker, 2, "");
+ABSL_FLAG(size_t, shared_log_conn_per_worker, 2, "");
 
-ABSL_FLAG(int, io_uring_entries, 128, "");
-ABSL_FLAG(int, io_uring_fd_slots, 128, "");
+ABSL_FLAG(size_t, io_uring_entries, 128, "");
+ABSL_FLAG(size_t, io_uring_fd_slots, 128, "");
 ABSL_FLAG(bool, io_uring_sqpoll, false, "");
 ABSL_FLAG(int, io_uring_sq_thread_idle_ms, 1, "");
 ABSL_FLAG(int, io_uring_cq_nr_wait, 1, "");
